Move `HackrStoryLoaderCacheDecorator` to production
