+++ HackrNewsApp/HackrNewsAppTests/Feed/HackrStoryLoaderCacheDecoratorTests.swift
@@ -3,28 +3,8 @@
 //
 
 import HackrNews
+@testable import HackrNewsApp
 import XCTest
-
-class HackrStoryLoaderCacheDecorator: HackrStoryLoader {
-    private let decoratee: HackrStoryLoader
-    private let cache: HackrStoryCache
-
-    init(decoratee: HackrStoryLoader, cache: HackrStoryCache) {
-        self.decoratee = decoratee
-        self.cache = cache
-    }
-
-    func load(id: Int, completion: @escaping (HackrStoryLoader.Result) -> Void) -> HackrStoryLoaderTask {
-        decoratee.load(id: id) { [unowned self] result in
-            switch result {
-            case let .success(story):
-                self.cache.save(story) { _ in }
-            case let .failure(error):
-                completion(.failure(error))
-            }
-        }
-    }
-}
 
 final class HackrStoryLoaderCacheDecoratorTests: XCTestCase {
     func test_init_doesNotLoadStoryData() {
